+++ lib/scaffold.dart
@@ -1,7 +1,7 @@
 library scaffold;
 
 export 'src/container/scaffold_container.dart';
-export 'src/routes/scaffold_routes.dart';
+export 'src/routes/routes.dart';
 export 'src/store/store.dart';
-export 'src/widgets/scaffold_state.dart';
+export 'src/widgets/injection_state.dart';
 export 'src/widgets/scoped_builder.dart';