+++ lib/src/routes/routes.dart
@@ -1,5 +1,5 @@
 import 'package:go_router/go_router.dart';
 
-abstract class ScaffoldRoutes {
+abstract class Routes {
   static List<GoRoute> get routes => const [];
 }