+++ lib/src/store/store.dart
@@ -2,8 +2,8 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 
 import '../state/abstract_state.dart';
 
-class ScaffoldStore<T> extends Cubit<AbstractState> {
-  ScaffoldStore() : super(InitialState());
+class Store<T> extends Cubit<AbstractState> {
+  Store() : super(InitialState());
 
   void execute(Future<T> Function() fn) async {
     try {